+++ ldap4j-java/src/main/java/hu/gds/ldap4j/ldap/AbandonRequest.java
@@ -4,13 +4,8 @@ import hu.gds.ldap4j.net.ByteBuffer;
 import org.jetbrains.annotations.NotNull;
 
 public record AbandonRequest(
-        int messageId,
-        boolean signKludge)
+        int messageId)
         implements Message<AbandonRequest> {
-    public AbandonRequest(int messageId) {
-        this(messageId, true);
-    }
-
     @Override
     public @NotNull AbandonRequest self() {
         return this;
@@ -20,6 +15,6 @@ public record AbandonRequest(
     public @NotNull ByteBuffer write() throws Throwable {
         return DER.writeTag(
                 Ldap.PROTOCOL_OP_ABANDON_REQUEST,
-                DER.writeIntegerNoTag(signKludge, messageId));
+                DER.writeIntegerNoTag(messageId));
     }
 }